+++ src/components/OverflowMenu/OverflowMenu_snapshot.spec.js
@@ -7,39 +7,26 @@ import OverflowMenu from './OverflowMenu';
 describe('OverflowMenu Component', () => {
   describe('snapshots', () => {
     it('should match snapshot', () => {
-    
       const wrapper1 = shallow(
         <OverflowMenu>
-          <OverflowMenu.Item onClick={() => alert("this.edit()")}>
-            Edit
-          </OverflowMenu.Item>
-          <OverflowMenu.Item onClick={() => alert("this.delete()")}>
-            Delete
-          </OverflowMenu.Item>
+          <OverflowMenu.Item onClick={() => alert('this.edit()')}>Edit</OverflowMenu.Item>
+          <OverflowMenu.Item onClick={() => alert('this.delete()')}>Delete</OverflowMenu.Item>
         </OverflowMenu>
       );
       expect(wrapper1).toMatchSnapshot();
 
       const wrapper2 = shallow(
         <OverflowMenu size="medium">
-          <OverflowMenu.Item onClick={() => alert("this.edit()")}>
-            Edit
-          </OverflowMenu.Item>
-          <OverflowMenu.Item onClick={() => alert("this.delete()")}>
-            Delete
-          </OverflowMenu.Item>
+          <OverflowMenu.Item onClick={() => alert('this.edit()')}>Edit</OverflowMenu.Item>
+          <OverflowMenu.Item onClick={() => alert('this.delete()')}>Delete</OverflowMenu.Item>
         </OverflowMenu>
       );
       expect(wrapper2).toMatchSnapshot();
 
       const wrapper3 = shallow(
         <OverflowMenu size="large">
-          <OverflowMenu.Item onClick={() => alert("this.edit()")}>
-            Edit
-          </OverflowMenu.Item>
-          <OverflowMenu.Item onClick={() => alert("this.delete()")}>
-            Delete
-          </OverflowMenu.Item>
+          <OverflowMenu.Item onClick={() => alert('this.edit()')}>Edit</OverflowMenu.Item>
+          <OverflowMenu.Item onClick={() => alert('this.delete()')}>Delete</OverflowMenu.Item>
         </OverflowMenu>
       );
       expect(wrapper3).toMatchSnapshot();
@@ -53,60 +40,40 @@ describe('OverflowMenu Component', () => {
 
       const wrapper5 = shallow(
         <OverflowMenu position="left">
-          <OverflowMenu.Item>
-            Edit
-          </OverflowMenu.Item>
-          <OverflowMenu.Item>
-            Delete
-          </OverflowMenu.Item>
+          <OverflowMenu.Item>Edit</OverflowMenu.Item>
+          <OverflowMenu.Item>Delete</OverflowMenu.Item>
         </OverflowMenu>
       );
       expect(wrapper5).toMatchSnapshot();
 
       const wrapper6 = shallow(
         <OverflowMenu appearance="primary">
-          <OverflowMenu.Item>
-            Edit
-          </OverflowMenu.Item>
-          <OverflowMenu.Item>
-            Delete
-          </OverflowMenu.Item>
+          <OverflowMenu.Item>Edit</OverflowMenu.Item>
+          <OverflowMenu.Item>Delete</OverflowMenu.Item>
         </OverflowMenu>
       );
       expect(wrapper6).toMatchSnapshot();
 
       const wrapper7 = shallow(
-        <OverflowMenu appearance="secondary">
-          <OverflowMenu.Item>
-            Edit
-          </OverflowMenu.Item>
-          <OverflowMenu.Item>
-            Delete
-          </OverflowMenu.Item>
+        <OverflowMenu appearance="danger">
+          <OverflowMenu.Item>Edit</OverflowMenu.Item>
+          <OverflowMenu.Item>Delete</OverflowMenu.Item>
         </OverflowMenu>
       );
       expect(wrapper7).toMatchSnapshot();
 
       const wrapper8 = shallow(
-        <OverflowMenu color='red'>
-          <OverflowMenu.Item>
-            Action 1
-          </OverflowMenu.Item>
-          <OverflowMenu.Item>
-            Action 2
-          </OverflowMenu.Item>
+        <OverflowMenu color="red">
+          <OverflowMenu.Item>Action 1</OverflowMenu.Item>
+          <OverflowMenu.Item>Action 2</OverflowMenu.Item>
         </OverflowMenu>
       );
       expect(wrapper8).toMatchSnapshot();
 
       const wrapper9 = shallow(
         <OverflowMenu activeColor="blue">
-          <OverflowMenu.Item>
-            Action 1
-          </OverflowMenu.Item>
-          <OverflowMenu.Item>
-            Action 2
-          </OverflowMenu.Item>
+          <OverflowMenu.Item>Action 1</OverflowMenu.Item>
+          <OverflowMenu.Item>Action 2</OverflowMenu.Item>
         </OverflowMenu>
       );
       expect(wrapper9).toMatchSnapshot();
@@ -114,16 +81,11 @@ describe('OverflowMenu Component', () => {
       const wrapper10 = shallow(
         <OverflowMenu open={true}>
           <OverflowMenu.Title>You can't close me</OverflowMenu.Title>
-          <OverflowMenu.Item>
-            Action 1
-          </OverflowMenu.Item>
-          <OverflowMenu.Item>
-            Action 2
-          </OverflowMenu.Item>
+          <OverflowMenu.Item>Action 1</OverflowMenu.Item>
+          <OverflowMenu.Item>Action 2</OverflowMenu.Item>
         </OverflowMenu>
       );
       expect(wrapper10).toMatchSnapshot();
-
     });
   });
 });